+++ CHANGELOG.md
@@ -9,7 +9,7 @@
   - a setting for the debouncing gap
 - some refactor on settings
 - setting for starting the language server eagerly or not
-- setting for file progressing all open editors eagerly or not (requires lsp 0.8.0)
+- setting for file progressing all open editors eagerly or not (requires lsp4ij 0.8.0)
 
 ## [0.1.5] - 2024-11-17